Move new IL patch methods and attribute declarations to respective classes
