+++ Celeste.Mod.mm/MonoModRules.cs
@@ -72,43 +72,6 @@ namespace MonoMod {
     [MonoModCustomMethodAttribute(nameof(MonoModRules.PatchInitblk))]
     class PatchInitblkAttribute : Attribute { }
 
-    /// <summary>
-    /// Patches the method to support non-looping complete screen layers.
-    /// </summary>
-    [MonoModCustomMethodAttribute(nameof(MonoModRules.PatchCompleteRendererImageLayerRender))]
-    class PatchCompleteRendererImageLayerRenderAttribute : Attribute { }
-
-    /// <summary>
-    /// Patches the method to update the Name and the TheoSisterName, if the file has been renamed in-game, in the file's SaveData.
-    /// </summary>
-    [MonoModCustomMethodAttribute(nameof(MonoModRules.PatchOuiFileSelectSlotOnContinueSelected))]
-    class PatchOuiFileSelectSlotOnContinueSelectedAttribute : Attribute { }
-    
-    /// <summary>
-    /// Patches MoveBlock.Controller to disable static movers before resetting their position when breaking.
-    /// </summary>
-    [MonoModCustomMethodAttribute(nameof(MonoModRules.PatchMoveBlockController))]
-    class PatchMoveBlockControllerAttribute : Attribute { }
-
-    /// <summary>
-    /// Patches the method to un-hardcode the FMOD event string used to play the footstep/landing sound effect.
-    /// </summary>
-    [MonoModCustomMethodAttribute(nameof(MonoModRules.PatchPlayerOnCollideV))]
-    class PatchPlayerOnCollideV : Attribute { }
-
-    /// <summary>
-    /// Patches the method to un-hardcode the FMOD event string used to play the grab sound effect.
-    /// </summary>
-    [MonoModCustomMethodAttribute(nameof(MonoModRules.PatchPlayerClimbBegin))]
-    class PatchPlayerClimbBegin : Attribute { }
-
-    /// <summary>
-    /// Patches the method to un-hardcode the FMOD event string used to play the walljump sound effect.
-    /// </summary>
-    [MonoModCustomMethodAttribute(nameof(MonoModRules.PatchPlayerOrigWallJump))]
-    class PatchPlayerOrigWallJumpAttribute: Attribute { }
-
-
     static partial class MonoModRules {
 
         static bool IsCeleste;
@@ -407,180 +370,6 @@ namespace MonoMod {
             cursor.Remove();
         }
 
-        public static void PatchPlayerCtorOnFrameChange(ILContext context) {
-            ILCursor cursor = new ILCursor(context);
-            PatchPlaySurfaceIndex(cursor, "/footstep");
-            PatchPlaySurfaceIndex(cursor, "/handhold");
-        }
-
-        public static void PatchPlayerOnCollideV(ILContext context, CustomAttribute attrib) {
-            MethodDefinition m_SurfaceIndex_GetPathFromIndex = context.Module.GetType("Celeste.SurfaceIndex").FindMethod("System.String GetPathFromIndex(System.Int32)");
-            MethodReference m_String_Concat = MonoModRule.Modder.Module.ImportReference(
-                MonoModRule.Modder.FindType("System.String").Resolve()
-                    .FindMethod("System.String Concat(System.String,System.String)")
-            );
-
-            ILCursor cursor = new ILCursor(context);
-
-            /*  Move cursor to after IL_040d in
-                    // num2 = platformByPriority.GetLandSoundIndex(this)
-                    IL_040a: ldloc.s 4
-	                IL_040c: ldarg.0
-	                IL_040d: callvirt instance int32 Celeste.Platform::GetLandSoundIndex(class Monocle.Entity)
-	                IL_0412: stloc.s 5
-                and get the variable index of num2  
-            */
-            int loc_landSoundIdx = -1;
-            cursor.GotoNext(MoveType.After,
-                instr => instr.MatchCallvirt("Celeste.Platform", "System.Int32 GetLandSoundIndex(Monocle.Entity)"),
-                instr => instr.MatchStloc(out loc_landSoundIdx));
-
-            /*  Change
-                    Play((playFootstepOnLand > 0f) ? "event:/char/madeline/footstep" : "event:/char/madeline/landing", "surface_index", num2);
-                to
-                    Play(SurfaceIndex.GetPathFromIndex(num2) + ((playFootstepOnLand > 0f) ? "/footstep" : "/landing"), "surface_index", num2);
-            */
-            cursor.GotoNext(MoveType.AfterLabel, instr => instr.MatchLdarg(0), instr => instr.MatchLdfld("Celeste.Player", "playFootstepOnLand"));
-            cursor.Emit(OpCodes.Ldloc, loc_landSoundIdx);
-            cursor.Emit(OpCodes.Call, m_SurfaceIndex_GetPathFromIndex);
-            cursor.GotoNext(instr => instr.MatchLdstr("event:/char/madeline/landing"))
-                .Next.Operand = "/landing";
-            cursor.GotoNext(instr => instr.MatchLdstr("event:/char/madeline/footstep"))
-                .Next.Operand = "/footstep";
-            cursor.GotoNext(MoveType.AfterLabel, instr => instr.MatchLdstr("surface_index"));
-            cursor.Emit(OpCodes.Call, m_String_Concat);
-        }
-
-        public static void PatchPlayerClimbBegin(ILContext context, CustomAttribute attrib) {
-            PatchPlaySurfaceIndex(new ILCursor(context), "/grab");
-        }
-
-        public static void PatchPlayerOrigWallJump(ILContext context, CustomAttribute attrib) {
-            // Doesn't use PatchPlaySurfaceIndex because index, not platform, is stored in the local variable
-
-            MethodDefinition m_SurfaceIndex_GetPathFromIndex = context.Module.GetType("Celeste.SurfaceIndex").FindMethod("System.String GetPathFromIndex(System.Int32)");
-            MethodReference m_String_Concat = MonoModRule.Modder.Module.ImportReference(
-                MonoModRule.Modder.FindType("System.String").Resolve()
-                    .FindMethod("System.String Concat(System.String,System.String)")
-            );
-            
-            ILCursor cursor = new ILCursor(context);
-
-            /*  Change:
-                    Play("event:/char/madeline/landing", "surface_index", num);
-                to:
-                    Play(SurfaceIndex.GetPathFromIndex(num) + "/landing", "surface_index", num);
-            */
-            cursor.GotoNext(MoveType.AfterLabel, instr => instr.MatchLdstr("event:/char/madeline/landing"));
-            cursor.Emit(OpCodes.Ldloc_0);
-            cursor.Emit(OpCodes.Call, m_SurfaceIndex_GetPathFromIndex);
-            cursor.Emit(OpCodes.Ldstr, "/landing");
-            cursor.Emit(OpCodes.Call, m_String_Concat);
-            cursor.Remove();
-        }
-
-        public static void PatchNPCSetupSpriteSoundsOnFrameChange(ILContext context) {
-            PatchPlaySurfaceIndex(new ILCursor(context), "/footstep");
-        }
-
-        public static void PatchCompleteRendererImageLayerRender(ILContext context, CustomAttribute attrib) {
-            MethodReference m_ImageLayer_get_ImageIndex = context.Method.DeclaringType.FindProperty("ImageIndex").GetMethod;
-            
-            ILCursor cursor = new ILCursor(context);
-            
-            // change: MTexture mTexture = Images[(int)(Frame % (float)Images.Count)];
-            // to:     MTexture mTexture = Images[ImageIndex];
-            // for new property ImageIndex
-            cursor.GotoNext(MoveType.After, instr => instr.MatchLdfld("Celeste.CompleteRenderer/ImageLayer", "Images"));
-            cursor.Emit(OpCodes.Ldarg_0);
-            cursor.Emit(OpCodes.Callvirt, m_ImageLayer_get_ImageIndex);
-            cursor.RemoveRange(8);
-        }
-        
-        public static void PatchOuiFileSelectSlotOnContinueSelected(ILContext context, CustomAttribute attrib) {
-            FieldDefinition f_OuiFileSelectSlot_Name = context.Method.DeclaringType.FindField("Name");
-            FieldDefinition f_OuiFileSelectSlot_renamed = context.Method.DeclaringType.FindField("renamed");
-            FieldDefinition f_SaveData_Name = context.Module.GetType("Celeste.SaveData").Resolve().FindField("Name");
-            FieldDefinition f_SaveData_TheoSisterName = context.Module.GetType("Celeste.SaveData").Resolve().FindField("TheoSisterName");
-            MethodDefinition m_Dialog_Clean = context.Module.GetType("Celeste.Dialog").Resolve().FindMethod("System.String Clean(System.String,Celeste.Language)");
-            TypeDefinition t_String = MonoModRule.Modder.FindType("System.String").Resolve();
-            MethodReference m_String_IndexOf = MonoModRule.Modder.Module.ImportReference(t_String.FindMethod("System.Int32 IndexOf(System.String,System.StringComparison)"));
-
-            // Insert after SaveData.Start(SaveData, FileSlot)
-            ILCursor cursor = new ILCursor(context);
-            cursor.GotoNext(MoveType.After, instr => instr.MatchCall("Celeste.SaveData", "System.Void Start(Celeste.SaveData,System.Int32)"));
-
-            // if (renamed)
-            // {
-            //     SaveData.Instance.Name = Name;
-            //     SaveData.Instance.TheoSisterName = Dialog.Clean((Name.IndexOf(Dialog.Clean("THEO_SISTER_NAME"), StringComparison.InvariantCultureIgnoreCase) >= 0) ? "THEO_SISTER_ALT_NAME" : "THEO_SISTER_NAME");
-            // }
-            ILLabel renamedTarget = cursor.DefineLabel();
-            ILLabel altNameTarget = cursor.DefineLabel();
-            ILLabel defaultNameTarget = cursor.DefineLabel();
-
-            // if (renamed)
-            cursor.Emit(OpCodes.Ldarg_0);
-            cursor.Emit(OpCodes.Ldfld, f_OuiFileSelectSlot_renamed);
-            cursor.Emit(OpCodes.Brfalse_S, renamedTarget);
-
-            // Assign Name
-            cursor.Emit(cursor.Next.OpCode, cursor.Next.Operand); // ldsfld class Celeste.SaveData Celeste.SaveData::Instance
-            cursor.Emit(OpCodes.Ldarg_0);
-            cursor.Emit(OpCodes.Ldfld, f_OuiFileSelectSlot_Name);
-            cursor.Emit(OpCodes.Stfld, f_SaveData_Name);
-
-            // Assign TheoSisterName
-            cursor.Emit(cursor.Next.OpCode, cursor.Next.Operand); // ldsfld class Celeste.SaveData Celeste.SaveData::Instance
-            cursor.Emit(OpCodes.Ldarg_0);
-            cursor.Emit(OpCodes.Ldfld, f_OuiFileSelectSlot_Name);
-            cursor.Emit(OpCodes.Ldstr, "THEO_SISTER_NAME");
-            cursor.Emit(OpCodes.Ldnull);
-            cursor.Emit(OpCodes.Call, m_Dialog_Clean);
-            cursor.Emit(OpCodes.Ldc_I4_3);
-            cursor.Emit(OpCodes.Callvirt, m_String_IndexOf);
-            cursor.Emit(OpCodes.Ldc_I4_0);
-            cursor.Emit(OpCodes.Bge_S, altNameTarget);
-            cursor.Emit(OpCodes.Ldstr, "THEO_SISTER_NAME");
-            cursor.Emit(OpCodes.Br_S, defaultNameTarget);
-            cursor.MarkLabel(altNameTarget);
-            cursor.Emit(OpCodes.Ldstr, "THEO_SISTER_ALT_NAME");
-            cursor.MarkLabel(defaultNameTarget);
-            cursor.Emit(OpCodes.Ldnull);
-            cursor.Emit(OpCodes.Call, m_Dialog_Clean);
-            cursor.Emit(OpCodes.Stfld, f_SaveData_TheoSisterName);
-
-            // Target for if renamed is false
-            cursor.MarkLabel(renamedTarget);
-        }
-        
-        public static void PatchMoveBlockController(MethodDefinition method, CustomAttribute attrib) {
-            MethodReference m_Platform_MoveStaticMovers = MonoModRule.Modder.Module.GetType("Celeste.Platform").FindMethod("System.Void MoveStaticMovers(Microsoft.Xna.Framework.Vector2)");
-
-            // The routine is stored in a compiler-generated method.
-            foreach (TypeDefinition nest in method.DeclaringType.NestedTypes) {
-                if (!nest.Name.StartsWith("<" + method.Name + ">d__"))
-                    continue;
-                method = nest.FindMethod("System.Boolean MoveNext()") ?? method;
-                break;
-            }
-
-            // From:
-            //     this.MoveStaticMovers(this.startPosition - this.Position);
-            //     this.DisableStaticMovers();
-            // To:
-            //     Vector2 amount = this.startPosition - this.Position;
-            //     this.DisableStaticMovers();
-            //     this.MoveStaticMovers(amount);
-            ILCursor cursor = new ILCursor(new ILContext(method));
-            cursor.GotoNext(MoveType.Before, instr => instr.MatchCallvirt("Celeste.Platform", "MoveStaticMovers"));
-            cursor.Remove();
-            cursor.GotoNext(MoveType.After, instr => instr.MatchCallvirt("Celeste.Platform", "DisableStaticMovers"));
-
-            // The argument order happens to let us emit the two function calls adjacent to each other
-            cursor.Emit(OpCodes.Callvirt, m_Platform_MoveStaticMovers);
-        }
-
         /// <summary>
         /// Fix ILSpy unable to decompile enumerator after MonoMod patching<br />
         /// (<code>yield-return decompiler failed: Unexpected instruction in Iterator.Dispose()</code>)
@@ -600,16 +389,6 @@ namespace MonoMod {
         }
 
         public static void PostProcessor(MonoModder modder) {
-            // Patch Player::ctor's OnFrameChange delegate manually because see above
-            new ILContext(modder.Module.GetType("Celeste.Player").FindMethod("<.ctor>b__280_1"))
-                .Invoke(PatchPlayerCtorOnFrameChange);
-            // Patch NPC::SetupTheoSpriteSounds's OnFrameChange delegate manually because see above
-            // We can also re-use the same patch code for NPC::SetupGrannySpriteSounds's OnFrameChange delegate
-            new ILContext(modder.Module.GetType("Celeste.NPC").FindMethod("<SetupTheoSpriteSounds>b__19_0"))
-                .Invoke(PatchNPCSetupSpriteSoundsOnFrameChange);
-            new ILContext(modder.Module.GetType("Celeste.NPC").FindMethod("<SetupGrannySpriteSounds>b__20_0"))
-                .Invoke(PatchNPCSetupSpriteSoundsOnFrameChange);
-
             // Patch previously registered AreaCompleteCtors and LevelExitRoutines _in that order._
             foreach (MethodDefinition method in AreaCompleteCtors)
                 PatchAreaCompleteCtor(method);
