+++ Celeste.Mod.mm/Patches/OuiFileSelectSlot.cs
@@ -16,6 +16,7 @@ using System.Linq;
 using Mono.Cecil;
 using Mono.Cecil.Cil;
 using MonoMod.Cil;
+using MonoMod.InlineRT;
 using MonoMod.Utils;
 
 namespace Celeste {
@@ -326,6 +327,12 @@ namespace MonoMod {
     [MonoModCustomMethodAttribute(nameof(MonoModRules.PatchFileSelectSlotRender))]
     class PatchFileSelectSlotRenderAttribute : Attribute { }
 
+    /// <summary>
+    /// Patches the method to update the Name and the TheoSisterName, if the file has been renamed in-game, in the file's SaveData.
+    /// </summary>
+    [MonoModCustomMethodAttribute(nameof(MonoModRules.PatchOuiFileSelectSlotOnContinueSelected))]
+    class PatchOuiFileSelectSlotOnContinueSelectedAttribute : Attribute { }
+
     static partial class MonoModRules {
 
         public static void PatchFileSelectSlotRender(ILContext context, CustomAttribute attrib) {
@@ -397,6 +404,63 @@ namespace MonoMod {
             cursor.RemoveRange(8);
             // Replace with `this.farewellStamp`
             cursor.Emit(OpCodes.Ldfld, f_farewellStamp);
+        }
+
+        public static void PatchOuiFileSelectSlotOnContinueSelected(ILContext context, CustomAttribute attrib) {
+            FieldDefinition f_OuiFileSelectSlot_Name = context.Method.DeclaringType.FindField("Name");
+            FieldDefinition f_OuiFileSelectSlot_renamed = context.Method.DeclaringType.FindField("renamed");
+            FieldDefinition f_SaveData_Name = context.Module.GetType("Celeste.SaveData").Resolve().FindField("Name");
+            FieldDefinition f_SaveData_TheoSisterName = context.Module.GetType("Celeste.SaveData").Resolve().FindField("TheoSisterName");
+            MethodDefinition m_Dialog_Clean = context.Module.GetType("Celeste.Dialog").Resolve().FindMethod("System.String Clean(System.String,Celeste.Language)");
+            TypeDefinition t_String = MonoModRule.Modder.FindType("System.String").Resolve();
+            MethodReference m_String_IndexOf = MonoModRule.Modder.Module.ImportReference(t_String.FindMethod("System.Int32 IndexOf(System.String,System.StringComparison)"));
+
+            // Insert after SaveData.Start(SaveData, FileSlot)
+            ILCursor cursor = new ILCursor(context);
+            cursor.GotoNext(MoveType.After, instr => instr.MatchCall("Celeste.SaveData", "System.Void Start(Celeste.SaveData,System.Int32)"));
+
+            // if (renamed)
+            // {
+            //     SaveData.Instance.Name = Name;
+            //     SaveData.Instance.TheoSisterName = Dialog.Clean((Name.IndexOf(Dialog.Clean("THEO_SISTER_NAME"), StringComparison.InvariantCultureIgnoreCase) >= 0) ? "THEO_SISTER_ALT_NAME" : "THEO_SISTER_NAME");
+            // }
+            ILLabel renamedTarget = cursor.DefineLabel();
+            ILLabel altNameTarget = cursor.DefineLabel();
+            ILLabel defaultNameTarget = cursor.DefineLabel();
+
+            // if (renamed)
+            cursor.Emit(OpCodes.Ldarg_0);
+            cursor.Emit(OpCodes.Ldfld, f_OuiFileSelectSlot_renamed);
+            cursor.Emit(OpCodes.Brfalse_S, renamedTarget);
+
+            // Assign Name
+            cursor.Emit(cursor.Next.OpCode, cursor.Next.Operand); // ldsfld class Celeste.SaveData Celeste.SaveData::Instance
+            cursor.Emit(OpCodes.Ldarg_0);
+            cursor.Emit(OpCodes.Ldfld, f_OuiFileSelectSlot_Name);
+            cursor.Emit(OpCodes.Stfld, f_SaveData_Name);
+
+            // Assign TheoSisterName
+            cursor.Emit(cursor.Next.OpCode, cursor.Next.Operand); // ldsfld class Celeste.SaveData Celeste.SaveData::Instance
+            cursor.Emit(OpCodes.Ldarg_0);
+            cursor.Emit(OpCodes.Ldfld, f_OuiFileSelectSlot_Name);
+            cursor.Emit(OpCodes.Ldstr, "THEO_SISTER_NAME");
+            cursor.Emit(OpCodes.Ldnull);
+            cursor.Emit(OpCodes.Call, m_Dialog_Clean);
+            cursor.Emit(OpCodes.Ldc_I4_3);
+            cursor.Emit(OpCodes.Callvirt, m_String_IndexOf);
+            cursor.Emit(OpCodes.Ldc_I4_0);
+            cursor.Emit(OpCodes.Bge_S, altNameTarget);
+            cursor.Emit(OpCodes.Ldstr, "THEO_SISTER_NAME");
+            cursor.Emit(OpCodes.Br_S, defaultNameTarget);
+            cursor.MarkLabel(altNameTarget);
+            cursor.Emit(OpCodes.Ldstr, "THEO_SISTER_ALT_NAME");
+            cursor.MarkLabel(defaultNameTarget);
+            cursor.Emit(OpCodes.Ldnull);
+            cursor.Emit(OpCodes.Call, m_Dialog_Clean);
+            cursor.Emit(OpCodes.Stfld, f_SaveData_TheoSisterName);
+
+            // Target for if renamed is false
+            cursor.MarkLabel(renamedTarget);
         }
 
     }
